SHRINKRES-274 Added support for running maven builds as a daemon (#119)
It is possible to stop the execution of the main thread and wait till some line matches a regex.  Useful for cases when maven build triggers some server or other application (eg. using mvn spring-boot:run)

The built() method then returns a representation of DaemonBuild - currently, it only returns if the surrounding thread is alive and if not then it can also return an instance of BuiltProject class (as a representation of the built project)