Use Plexus container to grab Maven components
As one of the reasons to container is exactly to protect
you (client code) for breakage that is not gonna happen
with components, as instantiation is done by container.

This code is now fully compatible with ALL line of Maven
versions released so far (from lowest that shrinkwrap
resolver requires of course).